+++ src/com/ansorgit/plugins/bash/editor/codecompletion/BashCompletionContributor.java
@@ -35,9 +35,9 @@ public class BashCompletionContributor extends CompletionContributor {
     private final static TokenSet endTokens = TokenSet.create(BashTokenTypes.STRING_END, BashTokenTypes.RIGHT_CURLY);
     private final static TokenSet wordTokens = TokenSet.create(BashTokenTypes.WORD);
 
-    public BashCompletionContributor() {
+    public BashCompletionContributor(BashPathCompletionService completionService) {
         new VariableNameCompletionProvider().addTo(this);
-        new CommandNameCompletionProvider().addTo(this);
+        new CommandNameCompletionProvider(completionService).addTo(this);
         new AbsolutePathCompletionProvider().addTo(this);
         new ShebangPathCompletionProvider().addTo(this);
         new DynamicPathCompletionProvider().addTo(this);