+++ Changelog.md
@@ -1,8 +1,13 @@
+#### 2018-08-27:
+ - Add action to create a report on the currently detected file type
+
 #### 2018-08-29:
  - \#522 Not a valid identifier in shell script with Bash functions with dash
 
 #### 2018-08-25:
  - fix to improve handling of local variables in functions
+ - \#566: Fix slow startup of the IDE when BashSupport is installed
+ - Enhancement: show path next to path command completion items
 
 #### 2018-08-14:
  - new icons for plugin (contributed by nosovae-dev)